+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.11.1 / 5.66.1] - 2023-08-??
 
+### Added
+- added 'RamDiskLetter=R:\' option allowing to mount the ramdisk root to a drive letter
+
 ### Changed
 - changed the new optin layout to now be the default for non vintage views (can be changed back in the settings)